+++ lib/ricoh_theta.dart
@@ -5,10 +5,30 @@ import 'package:ricoh_theta/models/image_infoes.dart';
 import 'package:ricoh_theta/models/picture_take_info.dart';
 import 'package:ricoh_theta/models/ricoh_image.dart';
 import 'package:ricoh_theta/models/storage_info.dart';
+import 'package:wifi_iot/wifi_iot.dart';
 
 import 'ricoh_theta_platform_interface.dart';
 
 class RicohTheta {
+  Future<dynamic> _forceWifi(Function function) async {
+    final isConnectedToWifi = await WiFiForIoTPlugin.isConnected();
+    if (!isConnectedToWifi) {
+      throw Exception('not connected to wifi device');
+    }
+
+    await WiFiForIoTPlugin.forceWifiUsage(true);
+    return await function();
+  }
+
+  Future<bool> isConnectedToTheta() async {
+    final ssid = await WiFiForIoTPlugin.getSSID();
+    if (ssid == null) {
+      return false;
+    }
+
+    return ssid.toLowerCase().contains("theta");
+  }
+
   /// If no ip address is provided, the default ip "192.168.1.1" is selected.
   /// This is required to setup before any other method.
   Future setTargetIp(String? ipAddress) async {
@@ -17,70 +37,110 @@ class RicohTheta {
 
   /// Disconnect from the device
   Future disconnect() async {
-    return RicohThetaPlatform.instance.disconnect();
+    await _forceWifi(
+      () => RicohThetaPlatform.instance.disconnect(),
+    );
+    await WiFiForIoTPlugin.forceWifiUsage(false);
   }
 
   /// Start capture of live view
   Future startLiveView({int fps = 30}) async {
-    return RicohThetaPlatform.instance.startLiveView(fps);
+    return _forceWifi(
+      () => RicohThetaPlatform.instance.startLiveView(fps),
+    );
   }
 
   /// Remove an image from storage device.
   /// Return true if the image is removed successfully.
-  Future<bool?> removeImageWithFileId(String fileId) {
-    return RicohThetaPlatform.instance.removeImageWithFileId(fileId);
+  Future<bool?> removeImageWithFileId(String fileId) async {
+    bool? result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.removeImageWithFileId(fileId);
+    });
+    return result;
   }
 
   /// Get an image from storage device.
   Future<RicohImage?> getImage(String fileId, String path) async {
-    return RicohThetaPlatform.instance.getImage(fileId, path);
+    RicohImage? result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.getImage(fileId, path);
+    });
+    return result;
   }
 
   /// Pause capture of live view
   Future pauseLiveView() async {
-    return RicohThetaPlatform.instance.pauseLiveView();
+    return _forceWifi(
+      () => RicohThetaPlatform.instance.pauseLiveView(),
+    );
   }
 
   /// Stop capture of live view
   Future stopLiveView() async {
-    return RicohThetaPlatform.instance.stopLiveView();
+    return _forceWifi(
+      () => RicohThetaPlatform.instance.stopLiveView(),
+    );
   }
 
   /// Resume capture of live view
   Future resumeLiveView() async {
-    return RicohThetaPlatform.instance.resumeLiveView();
+    return _forceWifi(
+      () => RicohThetaPlatform.instance.resumeLiveView(),
+    );
   }
 
   /// Get battery level from device
   Future<num> batteryLevel() async {
-    final battery = await RicohThetaPlatform.instance.batteryLevel();
-    return ((battery ?? 0) * 100).floor();
+    num? result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.batteryLevel();
+    });
+    return ((result ?? 0) * 100).floor();
   }
 
   /// Returns the current model information like model, firmware & serial number.
-  Future<DeviceInfo?> getDeviceInfo() {
-    return RicohThetaPlatform.instance.getDeviceInfo();
+  Future<DeviceInfo?> getDeviceInfo() async {
+    DeviceInfo? result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.getDeviceInfo();
+    });
+    return result;
   }
 
   /// Returns information about the device storage.
-  Future<StorageInfo?> getStorageInfo() {
-    return RicohThetaPlatform.instance.getStorageInfo();
+  Future<StorageInfo?> getStorageInfo() async {
+    StorageInfo? result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.getStorageInfo();
+    });
+    return result;
   }
 
   /// Update device session.
   /// Can be used to keep a session alive.
   Future update() {
-    return RicohThetaPlatform.instance.update();
+    return _forceWifi(
+      () => RicohThetaPlatform.instance.update(),
+    );
   }
 
   /// Returns information about the images stored on the device.
-  Future<List<ImageInfoes>> getImageInfoes() {
-    return RicohThetaPlatform.instance.getImageInfoes();
+  Future<List<ImageInfoes>> getImageInfoes() async {
+    late List<ImageInfoes> result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.getImageInfoes();
+    });
+    return result;
   }
 
   /// Take a picture & return a thumbnail path.
   Future<PictureTakeInfo?> takePicture(String path) async {
-    return RicohThetaPlatform.instance.takePicture(path);
+    PictureTakeInfo? result;
+    await _forceWifi(() async {
+      result = await RicohThetaPlatform.instance.takePicture(path);
+    });
+    return result;
   }
 
   /// Listen for live preview images coming from device.
@@ -102,6 +162,8 @@ class RicohTheta {
   /// [...]
   /// - 0 signifies unlimited frames
   Future adjustLiveViewFps(int fps) {
-    return RicohThetaPlatform.instance.adjustLiveViewFps(fps);
+    return _forceWifi(
+      () => RicohThetaPlatform.instance.adjustLiveViewFps(fps),
+    );
   }
 }