+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Ensure to force WiFi connection.
+* Add method to check if connected to Theta WiFi.
+
 ## 1.0.0
 
 * Initial release.